backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
google-generativeai